+++ src/scripts/path.ts
@@ -1,13 +1,14 @@
 import {existsSync} from 'node:fs';
-import {join, resolve} from 'node:path';
 import {fileURLToPath} from 'node:url';
+
+import {join, resolve} from 'pathe';
 
 export const ROOT = resolve(fileURLToPath(import.meta.url), '../..');
 
 export const resolver = (path: string) => resolve(ROOT, path);
 
 const PROD_DIR = resolve(fileURLToPath(import.meta.url), '..');
-const PROD = existsSync(`${PROD_DIR}/components.json`);
+const PROD = existsSync(join(PROD_DIR, 'components.json'));
 
 export const CACHE_DIR = PROD
   ? resolve(`${PROD_DIR}/.nextui-cli-cache`)